+++ Assets/Scripts/Creature/AnimHelpers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ba8fae1c3efa6cf42aee13f8e27d905c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
