+++ Assets/Scripts/Creature/AnimHelpers/CreatureIdle.cs
@@ -5,12 +5,6 @@ using UnityEngine;
 public class CreatureIdle : StateMachineBehaviour
 {
     public int maxLoops = 3;
-
-    // OnStateEnter is called when a transition starts and the state machine starts to evaluate this state
-    override public void OnStateEnter(Animator animator, AnimatorStateInfo stateInfo, int layerIndex) {
-
-    }
-
     // OnStateUpdate is called on each Update frame between OnStateEnter and OnStateExit callbacks
     override public void OnStateUpdate(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
     {
