+++ Assets/Scripts/Creature/AnimHelpers/CreatureIdle.cs.meta
