+++ Assets/Scripts/Creature/AnimHelpers/CreatureRTurn.cs
@@ -1,0 +1,14 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class CreatureRTurn : StateMachineBehaviour
+{
+    // OnStateExit is called when a transition ends and the state machine finishes evaluating this state
+    override public void OnStateExit(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
+    {
+        animator.SetBool("idleLeftTurn", true);
+        animator.SetBool("idleRightTurn", false);
+        animator.SetBool("isLeft", true);
+    }
+}
