+++ Assets/Scripts/Creature/AnimHelpers/CreatureStunned.cs
@@ -1,0 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class CreatureStunned : StateMachineBehaviour
+{
+    // OnStateExit is called when a transition ends and the state machine finishes evaluating this state
+    override public void OnStateExit(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
+    {
+        CreatureManager.Instance.IsStunned = false;
+    }
+}
